requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pydub
moviepy
Pillow
python-multipart
requests
openai
librosa
soundfile
spotipy
numpy
scipy
gtts
python-dotenv
aubio
reportlab
PyJWT
passlib[bcrypt]
stripe
sqlalchemy
aiosqlite
pytest
pytest-asyncio
httpx
aiofiles
redis
